+++ test.sh
@@ -8,11 +8,19 @@ BLUE="\033[1;34m"
 SRC_DIR="./tests"
 OUT_DIR="./build"
 
+CC=cc
+CXX=c++
+
+COMMON_FLAGS="-Wall -Wextra -Werror -pedantic -Isrc -O0"
+CFLAGS="-std=c99"
+CXXFLAGS="-std=c++11"
+
 if [ -z "$CI" ] || [ "$RUNNER_OS" = "Linux" ]; then
-    FSANITIZE=-fsanitize=address,leak,undefined
+    FSANITIZE="-fsanitize=address,leak,undefined"
 else
     FSANITIZE=
 fi
+
 
 mkdir -p $OUT_DIR
 
@@ -21,9 +29,15 @@ failed=0
 for file in $(find $SRC_DIR -type f); do
     filename=$(basename $file)
     name=${filename%.*}
+    extension=${filename##*.}
     bin="$OUT_DIR/$name"
 
-    cc -Wall -Wextra -Werror -pedantic -std=c99 $FSANITIZE -Isrc $file -o $bin 2>/dev/null
+    if [ "$extension" = "c" ]; then
+        $CC $COMMON_FLAGS $FSANITIZE $CFLAGS $file -o $bin 2>/dev/null
+    else
+        $CXX $COMMON_FLAGS $FSANITIZE $CXXFLAGS $file -o $bin 2>/dev/null
+    fi
+
     if [ $? -ne 0 ]; then
         echo -e $RED"[COMPILATION FAILED] $name"$RESET
         ((failed += 1))